+++ Player/Player.cpp
@@ -222,7 +222,7 @@ bool Player::CheckCollisionWithBoss(XMFLOAT3 bossPos, float collisionRadius)
 	float dy = bossPos.y - position.y;
 	float dz = bossPos.z - position.z;
 
-	float distance = sqrtf(pow(dx, 2) + pow(dy, 2) + pow(dz, 2));
+	float distance = sqrtf(powf(dx, 2) + powf(dy, 2) + powf(dz, 2));
 
 	return distance <= collisionSphereRadius + collisionRadius;
 }
